Update client.cpp
small syntax fix for graphFile; 
output for measurementFile brings suspicion, output for T4 is always zero (This might just be an inherent fact from the problem which i'd need to think about ), on burst 1 sometimes there's empty entries of the form ,, or ,,, etc.. This is probably a formatting issue. For now can push forward to the plot and if the plot seems ridiculous can go back to fix and re-plot.